feat: Initialize admin dashboard with core layout, client-side authentication, and customer data table.
