+++ src/components/customers-dashboard/customers-table.tsx
@@ -118,7 +118,6 @@ const columns = [
     cell: (row: any) => <ActionsMenu row={row} />,
     ignoreRowClick: true,
     width: "60px",
-    right: true,
   },
 ];
